Refactor: 유효성 검증 스키마 분리
