+++ src/app/settings/components/EmailSetting/EmailSetting.tsx
@@ -11,7 +11,7 @@ import { MsgPlaceholder } from '@/assets/message';
 import { AiFillLock } from 'react-icons/ai';
 import { UserInfo } from '@/lib/types/auth';
 import { useForm } from 'react-hook-form';
-import { emailFormSchema } from '@/lib/validation/schame';
+import { emailFormSchema } from '@/lib/validation/userInfoSchema';
 import { yupResolver } from '@hookform/resolvers/yup';
 import useChangeUserInfo from '@/query/userinfo/useChangeUserInfo';
 import { handleCompareTwice } from '@/lib/handler/handleCompare';
